(#40) Clean. Organise feature and reducer setup classes
