+++ Packages/Fluxity/Runtime/Extensions/FlumeServiceContainerExt.cs
@@ -15,51 +15,4 @@ namespace AIR.Fluxity
             return self;
         }
     }
-
-    public class FluxityFlumeRegisterContext
-    {
-        public FluxityFlumeRegisterContext(Store store, Dispatcher distpatcher, FlumeServiceContainer flumeServiceContainer)
-        {
-            Store = store;
-            Distpatcher = distpatcher;
-            FlumeServiceContainer = flumeServiceContainer;
-        }
-
-        public Store Store { get; }
-        public Dispatcher Distpatcher { get; }
-        public FlumeServiceContainer FlumeServiceContainer { get; }
-
-        public FluxityFeatureContext<TState> Feature<TState>(TState startingValue)
-            where TState : struct
-        {
-            var feat = new Feature<TState>(startingValue);
-            Store.AddFeature(feat);
-            return new FluxityFeatureContext<TState>(this, feat);
-        }
-    }
-
-    public class FluxityFeatureContext<TState> : FluxityFlumeRegisterContext
-        where TState : struct
-    {
-        private readonly Feature<TState> _feature;
-        public FluxityFeatureContext(FluxityFlumeRegisterContext context, Feature<TState> feature)
-            : base(context.Store, context.Distpatcher, context.FlumeServiceContainer)
-        {
-            _feature = feature;
-        }
-
-        public FluxityFeatureContext<TState> BulkReducers(Action<FluxityFeatureContext<TState>> registerAll)
-        {
-            registerAll(this);
-            return this;
-        }
-
-        public FluxityFeatureContext<TState> Reducer<TCommand>(IReducer<TState, TCommand>.ReduceDelegate pureFunctionReducer)
-            where TCommand : ICommand
-        {
-            var reducer = new PureFunctionReducerBinder<TState, TCommand>(pureFunctionReducer);
-            _feature.Register(reducer);
-            return this;
-        }
-    }
 }
