+++ Packages/Fluxity/Runtime/FluxityFlumeRegisterContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70aa07304b9929b4d94ba57af0777c1d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
